integration: extract use_keyspace.rs
